python/requirements.txt
census==0.8.17
certifi==2021.5.30
charset-normalizer==2.0.3
idna==3.2
jellyfish==0.6.1
psycopg2-binary==2.9.1
requests==2.26.0
urllib3==1.26.6
us==2.0.2
